Turning ON the mutual TLS
This commit sets the mutual TLS option to ON by default.

Core mTLS implementation was accepted under this commit:

   https://gerrit.openbmc-project.xyz/c/openbmc/bmcweb/+/23588

where by default the mTLS was disabled.

Tested:
   Manual tests were made to verify if this option turns the
   mTLS ON.

+++ http/http_connection.h
@@ -275,7 +275,7 @@ class Connection
             SSL_set_session_id_context(
                 adaptor.native_handle(),
                 reinterpret_cast<const unsigned char*>(serverName.c_str()),
-                serverName.length());
+                static_cast<unsigned int>(serverName.length()));
             BMCWEB_LOG_DEBUG << this << " TLS is enabled on this connection.";
         }
